+++ ingress-controller/internal/controllers/gateway/route_predicate_test.go
@@ -196,7 +196,7 @@ func TestIsRouteAttachedToReconciledGateway(t *testing.T) {
 					},
 				},
 			},
-			expectedResult: false,
+			expectedResult: true,
 		},
 	}